Updated UI solution to use keywords
Took my initial (but complete) solution which was the steps to perform
my test and formed into a more keyword driven test. This is my normal
process for developing scripts. I'll start off with the steps I think
which will get me to the test ing I want and then I will go back and
refine that into a more keyword based script thinkking about the larger
design of the test suite.